+++ setup.py
@@ -1,0 +1,10 @@
+import setuptools
+
+setuptools.setup(
+        name="EXP_tools",
+        version="0.1",
+        author="B-BFE collaboration",
+        author_email="ngaravito@flatironinstitute.org",
+        description="Analysis tools for EXP and other BFE packages",
+        packages=["EXP_tools"],
+        )
